% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_interaction}
\alias{append_interaction}
\title{Append interaction}
\usage{
append_interaction(x, covariate, predictor)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{predictor}{character name}
}
\value{
numeric array
}
\description{
Append an interaction variable to a 4D array.
}
\examples{
x <- append_interaction(x, "temperature", "precipitation")

}
