% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_lags}
\alias{append_lags}
\title{Append lagged covariates}
\usage{
append_lags(x, covariate, lags)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{lags}{integers}
}
\value{
numeric array
}
\description{
Append lagged variables to a 4D array.
}
\examples{
x <- append_lags(x, "precipitation", 2)
x <- append_lags(x, "precipitation", c(2, 4, 6))

}
