% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_leading_terms}
\alias{append_leading_terms}
\title{Append leading covariates}
\usage{
append_leading_terms(x, covariate, leads)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{leads}{integers}
}
\value{
numeric array
}
\description{
Append leading variables to a 4D array.
}
