% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_pcs.R
\name{append_pcs}
\alias{append_pcs}
\title{Append Principal Components}
\usage{
append_pcs(x, covariates, npc, prefix)
}
\arguments{
\item{x}{numeric array}

\item{covariates}{vector of covariates}

\item{npc}{integer: # of principal components}

\item{prefix}{character prefix for column name}
}
\value{
numeric array
}
\description{
Append principal components to a 4D array.
Plots from \code{factoextra} describe the PCA.
}
