% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_principal_components}
\alias{append_principal_components}
\title{Append principal components}
\usage{
append_principal_components(x, covariates, npc, prefix)
}
\arguments{
\item{x}{numeric array}

\item{covariates}{vector of covariates}

\item{npc}{integer: number of principal components}

\item{prefix}{character prefix for column name}
}
\value{
numeric array
}
\description{
Append principal components to a 4D array.
Plots from \code{factoextra} describe the PCA.
}
\examples{
x <- append_principal_components(x, dimnames(x)[[4]][12:23], 2, "climate")

}
