% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_sp.R
\name{append_sp}
\alias{append_sp}
\title{Append Spatial Random Effects}
\usage{
append_sp(x, arealeffs, A)
}
\arguments{
\item{x}{numeric array}

\item{arealeffs}{character vector: areal effects}

\item{A}{adjacency matrix}
}
\value{
numeric array
}
\description{
Append eigenvectors from Moran operator decomposition
to a 4D array.
}
