% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_spatial_effects}
\alias{append_spatial_effects}
\title{Append spatial random effects}
\usage{
append_spatial_effects(x, arealeffs, A)
}
\arguments{
\item{x}{numeric array}

\item{arealeffs}{character vector: areal effects}

\item{A}{adjacency matrix}
}
\value{
numeric array
}
\description{
Append eigenvectors from Moran operator decomposition
to a 4D array.
}
\examples{
x <- append_spatial_effects(x, c("urban", "agri", "forest"), adjacency_matrix)

}
