% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_transform.R
\name{append_transform}
\alias{append_transform}
\title{Append Transformed Variable}
\usage{
append_transform(x, covariate, fn, prefix)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{fn}{simple function}

\item{prefix}{character prefix for column name}
}
\value{
numeric array
}
\description{
Append a transformed variable to a 4D array.
}
