% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvoccu_chain.R
\name{btvoccu_chain}
\alias{btvoccu_chain}
\title{Chaining 'btvoccu' models}
\usage{
btvoccu_chain(model1, model2)
}
\arguments{
\item{model1}{\code{btvoccu} model object}

\item{model2}{\code{btvoccu} model object}
}
\value{
\code{btvoccu} model object
}
\description{
Concatenate independent MCMC chains for \code{btvoccu} model objects
}
