% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvoccu_simstudy.R
\name{btvoccu_simstudy}
\alias{btvoccu_simstudy}
\title{Simulate time-varying occupancy model}
\usage{
btvoccu_simstudy(
  nsites,
  nseasons,
  nperiods,
  beta,
  alpha,
  occutime,
  dettime,
  naprob = 0,
  noise = 1,
  probit = F
)
}
\arguments{
\item{nsites}{number of sites}

\item{nseasons}{number of seasons}

\item{nperiods}{number of periods}

\item{beta}{vector of occupancy effects}

\item{alpha}{vector of detection effects}

\item{occutime}{array of time-varying occupancy covariates}

\item{dettime}{array of time-varying detection covariates}

\item{naprob}{percent of NAs in simulated data}

\item{noise}{numeric parameter controlling covariate noise}

\item{probit}{logical: default \code{F} for logit link and \code{T} for probit link}
}
\value{
list including simulated data and more
}
\description{
This function simulates data for a time-varying occupancy model.
}
