% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{combine_chains}
\alias{combine_chains}
\title{Combine MCMC chains}
\usage{
combine_chains(model, effect, mnridx, burnin = 0.5, thin = 1)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{effect}{"betas", "alphas", or "thetas"}

\item{mnridx}{integer index}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}
}
\value{
vector of posterior samples
}
\description{
Combine chains for mixing diagnostics
and posterior inference.
}
