% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{plot_btvoccu}
\alias{plot_btvoccu}
\title{\code{btvoccu} predictions visualized}
\usage{
plot_btvoccu(
  model,
  x,
  sites,
  season,
  periods,
  value = "Occupancy",
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975),
  ndraws = 1000,
  ribbons = TRUE,
  dots = FALSE,
  spline = FALSE,
  spline.predict = FALSE,
  nknots = 5,
  xaxis = "Period",
  legendtitle = "Site"
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric array: covariate array}

\item{sites}{vector}

\item{season}{length \code{1} vector}

\item{periods}{vector}

\item{value}{'Occupancy', 'Detection', or 'Presence'}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}

\item{credible}{quantiles (length \code{3})}

\item{ndraws}{number of posterior draws}

\item{ribbons}{logical: default \code{TRUE} to show uncertainty ribbons}

\item{dots}{logical: default \code{FALSE} to not show actual datapoints}

\item{spline}{logical: default \code{FALSE} to not fit smoothing splines}

\item{spline.predict}{logical: default \code{FALSE} to not predict missing values using splines}

\item{nknots}{see \code{smooth.spline()}}

\item{xaxis}{character: x-axis label}

\item{legendtile}{character: legend label}
}
\value{
\code{ggplot} object
}
\description{
Visualize occupancy, detection, and presence
probabilities with uncertainty over time.
}
