% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plots}
\usage{
plot_trace(model, effects, mnridx, burnin = 0)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{effects}{"betas", "alphas", or "thetas"}

\item{mnridx}{integer index}

\item{burnin}{percent of posterior samples to burn}
}
\description{
Plot trace of posterior samples.
}
\examples{
trace_plot(mdl, "betas", 1)
trace_plot(mdl, "alphas", 1, 0.10)
for(i in 1:length(mdl$occueffs)){plot_trace(mdl, "betaS", i, burnin = 0)}

}
