% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_check.R
\name{posterior_check}
\alias{posterior_check}
\title{Posterior simulation}
\usage{
posterior_check(
  model,
  x,
  y = NULL,
  sites = NULL,
  seasons = NULL,
  periods = NULL,
  nrep = 1000,
  ndraws = 1000,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric array: covariate array}

\item{y}{binary array: response array (default \code{NULL} does not provide posterior predictive checks)}

\item{sites}{vector (default \code{NULL} uses training sites)}

\item{seasons}{vector (default \code{NULL} uses training seasons)}

\item{periods}{vector (default \code{NULL} uses training periods)}

\item{nrep}{number of replicates}

\item{ndraws}{number of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth draw}
}
\value{
list
}
\description{
Simulate presence-absences based on a \code{btvoccu} model.
Output some predictive checks and statistics,
including a histogram.
}
\examples{
output <- posterior_check(mdl, x)
output <- posterior_check(mdl, x, NULL, c("TOR", "YRK", "PEE"), 2004:2005, 20:40)
output <- posterior_check(mdl, x, y, c("TOR", "YRK", "PEE"), 2004:2005, 20:40)

}
