% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_correlations.R
\name{posterior_correlations}
\alias{posterior_correlations}
\title{Posterior correlations}
\usage{
posterior_correlations(
  model,
  Sigmas,
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{Sigmas}{"Sigmabetas" or "Sigmaalphas"}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}

\item{credible}{quantiles}
}
\value{
matrix
}
\description{
Summarize posterior correlations for beta and alpha coefficients.
}
\examples{
View(posterior_correlations(mdl, "Sigmabetas"))
xtable(posterior_correlations(mdl, "Sigmaalphas", burnin = 0.10))

}
