% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_corrs.R
\name{posterior_corrs}
\alias{posterior_corrs}
\title{Posterior Correlations}
\usage{
posterior_corrs(
  model,
  Sigmas,
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{Sigmas}{"Sigmabetas" or "Sigmaalphas"}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}

\item{credible}{vector of quantiles}
}
\value{
matrix
}
\description{
Summarize posterior correlations.
}
