% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_effects.R
\name{posterior_effects}
\alias{posterior_effects}
\title{Posterior Effects}
\usage{
posterior_effects(
  model,
  credible = c(0.025, 0.5, 0.975),
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{credible}{vector of quantiles}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
matrix
}
\description{
Summarize posterior effects.
}
