% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sigmoids.R
\name{posterior_sigmoids}
\alias{posterior_sigmoids}
\title{Posterior Sigmoids}
\usage{
posterior_sigmoids(
  model,
  x,
  sites,
  seasons,
  periods,
  ndraws = 1000,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{x}{numeric array: predictors array}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}

\item{ndraws}{integer: # of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
numeric array
}
\description{
Compute posterior occupancy, detection,
and presence probabilities.
}
