% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_simulate.R
\name{posterior_simulate}
\alias{posterior_simulate}
\title{Posterior Simulation}
\usage{
posterior_simulate(
  model,
  x,
  y = NULL,
  sites = NULL,
  seasons = NULL,
  periods = NULL,
  nrep = 1000,
  ndraws = 1000,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{x}{numeric array: predictors array}

\item{y}{binary array: response array (default \code{NULL} does not provide posterior predictive checks)}

\item{sites}{vector (default \code{NULL} uses training sites)}

\item{seasons}{vector (default \code{NULL} uses training seasons)}

\item{periods}{vector (default \code{NULL} uses training periods)}

\item{nrep}{integer: # of replicates}

\item{ndraws}{integer: # of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
list
}
\description{
Simulate presence-absences based on a \code{btvoccu} model.
Output some predictive checks and statistics,
including a histogram.
}
