% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_vars.R
\name{posterior_vars}
\alias{posterior_vars}
\title{Posterior Variances}
\usage{
posterior_vars(
  model,
  Sigmas,
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{Sigmas}{"Sigmabetas", "Sigmaalphas", or "sigmathetas"}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}

\item{credible}{vector of quantiles}
}
\value{
matrix
}
\description{
Summarize posterior variances.
}
