% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ry.R
\name{ry}
\alias{ry}
\title{Simulate Presence-Absences}
\usage{
ry(pre)
}
\arguments{
\item{pre}{numeric array: presence probabilities}
}
\value{
binary array
}
\description{
Simulate presence-absence data
based on presence probabilities.
}
