% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{split_4darray}
\alias{split_4darray}
\title{Train-validate-test for 4D array}
\usage{
split_4darray(y, p, sites = NULL, seasons = NULL, periods = NULL)
}
\arguments{
\item{y}{binary array: response array}

\item{p}{proportions (vector length \code{1} or \code{2})}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}
}
\value{
list
}
\description{
Split 4D array into training,
validation, and testing arrays.
Splitting is done by setting
values to NA because main function
\code{btvoccu()} ignores NA responses
when fitting.
}
\examples{
ytest <- subset_4darray(y, 2, c(2008, 2012, 2016))
ysplit <- subset_4darray(y, 2, c(2002:2007, 2009:2011, 2013:2015, 2017))
ysplit <- split_4darray(y, .75, periods = 20:40)
ytrain <- ysplit$train
yvalid <- ysplit$test

ysplit <- split_4darray(y, c(6, .2), periods = 20:40)

}
