% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{subset_4darray}
\alias{subset_4darray}
\title{Subset 4D array}
\usage{
subset_4darray(x, d, labels)
}
\arguments{
\item{x}{numeric array}

\item{d}{integer: array dimension}

\item{labels}{character vector: predictor names}
}
\value{
numeric array
}
\description{
Subset 4-dimensional arrays along any major dimension.
}
\examples{
w <- subset_4darray(x, 1, c("TOR", "YRK", "PEE"))
w <- subset_4darray(x, 2, 2005:2010)
w <- subset_4darray(x, 3, 20:40)
w <- subset_4darray(x, 4, c("precipitation", "temperature", "popdensity"))

}
