% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Trace Plots}
\usage{
trace_plot(model, effects, mnridx, burnin = 0.5)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{effects}{"betas", "alphas", or "thetas"}

\item{mnridx}{integer index}

\item{burnin}{percent of posterior samples to burn}
}
\description{
Plot trace of posterior samples.
}
