% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_covariate_plot.R
\name{ts_covariate_plot}
\alias{ts_covariate_plot}
\title{Covariate Time Series Plots}
\usage{
ts_covariate_plot(
  x,
  sites,
  season,
  periods,
  covariate,
  yaxis,
  xaxis = "Epiweek",
  legendtitle = "Locality"
)
}
\arguments{
\item{x}{numeric array: predictors array}

\item{sites}{vector}

\item{season}{length \code{1} vector}

\item{periods}{vector}

\item{covariate}{character}

\item{yaxis}{characer: y-axis label}

\item{xaxis}{character: x-axis label}

\item{legendtile}{character: legend label}
}
\value{
\code{ggplot} object
}
\description{
Visualize covariates over time.
}
