% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predict.R
\name{ts_predict}
\alias{ts_predict}
\title{Predictive Time Series}
\usage{
ts_predict(
  model,
  x,
  sites,
  season,
  periods,
  credible = c(0.025, 0.5, 0.975),
  value = "Occupancy",
  spline = TRUE,
  spline.predict = FALSE,
  nknots = 5,
  ndraws = 1000,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{x}{numeric array: predictors array}

\item{sites}{character vector}

\item{season}{integer}

\item{periods}{integer vector}

\item{credible}{length \code{3} vector of quantiles}

\item{value}{'Occupancy', 'Detection', or 'Presence'}

\item{spline}{logical: default \code{TRUE} to fit smoothing splines}

\item{spline.predict}{logical: default \code{FALSE} to not predict missing values using splines}

\item{nknots}{see \code{smooth.spline()}}

\item{ndraws}{integer: # of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
data frame
}
\description{
Predict occupancy, detection, and presence
probabilities with uncertainty over time.
}
