% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predict_plot.R
\name{ts_predict_plot}
\alias{ts_predict_plot}
\title{Predictive Time Series Plots}
\usage{
ts_predict_plot(
  model,
  x,
  sites,
  season,
  periods,
  credible = c(0.025, 0.5, 0.975),
  value = "Occupancy",
  ribbons = TRUE,
  spline = FALSE,
  spline.predict = FALSE,
  nknots = 5,
  xaxis = "Epiweek",
  legendtitle = "Locality",
  dots = FALSE,
  ndraws = 1000,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{x}{numeric array: predictors array}

\item{sites}{vector}

\item{season}{length \code{1} vector}

\item{periods}{vector}

\item{credible}{length \code{3} vector of quantiles}

\item{value}{'Occupancy', 'Detection', or 'Presence'}

\item{ribbons}{logical: default \code{TRUE} to show uncertainty ribbons}

\item{spline}{logical: default \code{FALSE} to not fit smoothing splines}

\item{spline.predict}{logical: default \code{FALSE} to not predict missing values using splines}

\item{nknots}{see \code{smooth.spline()}}

\item{xaxis}{character: x-axis label}

\item{dots}{logical: default \code{FALSE} to not show actual datapoints}

\item{ndraws}{integer: # of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}

\item{legendtile}{character: legend label}
}
\value{
\code{ggplot} object
}
\description{
Visualize occupancy, detection, and presence
probabilities with uncertainty over time.
}
