% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{waic_score}
\alias{waic_score}
\title{Watanabe Akaike information criterion}
\usage{
waic_score(
  model,
  x,
  y,
  sites = NULL,
  seasons = NULL,
  periods = NULL,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric array: covariate array}

\item{y}{binary array: response array}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}
}
\value{
list
}
\description{
Compute Watanabe-Akaike information criterion.
When indices are set to \code{NULL},
WAIC is calculated using the training data.
}
