% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_interaction_term.R
\name{append_interaction_term}
\alias{append_interaction_term}
\title{Append interaction term}
\usage{
append_interaction_term(
  x,
  covariate,
  predictor,
  fcn = function(a, b) {     a * b },
  fname = "*"
)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{predictor}{character name}

\item{fcn}{function, default is to multiply}

\item{fname}{function name}
}
\value{
numeric array
}
\description{
Append an interaction variable to a 4D array.
}
\examples{
x <- append_interaction_term(x, "temperature", "precipitation")

}
