% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_lags.R
\name{append_lags}
\alias{append_lags}
\title{Append Lagged Covariates}
\usage{
append_lags(x, covariate, lags)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{lags}{integers}
}
\value{
numeric array
}
\description{
Append lagged variables to a 4D array.
}
