% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4darray.R
\name{append_transform}
\alias{append_transform}
\title{Append transformed variable}
\usage{
append_transform(x, covariate, fn, prefix)
}
\arguments{
\item{x}{numeric array}

\item{covariate}{character name}

\item{fn}{function}

\item{prefix}{character prefix for column name}
}
\value{
numeric array
}
\description{
Append a transformed variable to a 4D array.
}
\examples{
x <- append_transform(x, "precipitation", sqrt, "sqrt.")
x <- append_transform(x, "popdensity", log10, "log10.")

}
