% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvoccu.R
\name{btvoccu}
\alias{btvoccu}
\title{Bayesian time-varying occupancy model}
\usage{
btvoccu(
  niter,
  x,
  y,
  sites,
  seasons,
  periods,
  occueffs,
  deteffs,
  speffs = NULL,
  A = NULL,
  logit = TRUE,
  nchains = 1,
  print.interval = 1000
)
}
\arguments{
\item{niter}{integer: # of iterations}

\item{x}{numeric array: covariate array}

\item{y}{binary array: response array}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}

\item{occueffs}{character vector: occupancy effects}

\item{deteffs}{character vector: detection effects}

\item{speffs}{character vector: spatial random effects}

\item{A}{adjacency matrix}

\item{logit}{logical: default \code{TRUE} for logistic regression and \code{FALSE} for probit regression}

\item{nchains}{number of chains (default \code{1})}

\item{print.interval}{integer: period to print iteration to console (default \code{1000})}
}
\value{
length \code{15} list of posterior samples and more
}
\description{
Train a time-varying occupancy model
using Bayesian generalized linear regression.
}
