% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvoccu_map.R
\name{btvoccu_map}
\alias{btvoccu_map}
\title{Thematic \code{btvoccu} Maps}
\usage{
btvoccu_map(
  file,
  model,
  x,
  value,
  q,
  shape,
  cities,
  sites,
  season,
  period,
  palette = "Reds",
  legend.position = c("left", "bottom"),
  title.position = c("right", "top"),
  compass.position = c("right", "bottom"),
  title.size = 0.8,
  legend.title.size = 0.8,
  legend.text.size = 0.6
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric covariate array}

\item{value}{"Occupancy", "Detection", or "Presence"}

\item{q}{quantile}

\item{shape}{\code{SpatialPolygonsDataFrame} object}

\item{cities}{\code{sf} object, named points}

\item{sites}{vector (in order of \code{shape} object)}

\item{season}{integer}

\item{period}{integer}

\item{palette}{color palette}

\item{legend.position}{see \code{tmap::tm_layout()}}

\item{title.position}{see \code{tmap::tm_layout()}}

\item{compass.position}{see \code{tmap::tm_compass()}}

\item{title.size}{see \code{tmap::tm_layout()}}

\item{legend.title.size}{see \code{tmap::tm_layout()}}

\item{legend.text.size}{see \code{tmap::tm_layout()}}
}
\value{
\code{tmap} object
}
\description{
Create thematic map demonstrating how
output sigmoids vary across space.
}
