% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btvoccu_movie.R
\name{btvoccu_movie}
\alias{btvoccu_movie}
\title{Thematic \code{btvoccu} Maps Animated}
\usage{
btvoccu_movie(
  file,
  directory,
  fps,
  model,
  x,
  value,
  q,
  shape,
  cities,
  sites,
  season,
  period,
  palette = "Reds",
  legend.position = c("left", "top"),
  title.position = c("right", "top"),
  title.size = 0.8,
  legend.title.size = 0.8,
  legend.text.size = 0.6,
  xlab.size = 0.8,
  ylab.size = 0.8,
  dots.size = 0.04,
  cities.text = "name",
  cities.just = "top",
  cities.size = 0.7,
  cities.col = "black"
)
}
\arguments{
\item{file}{output file name}

\item{directory}{output directory name (must exist already)}

\item{fps}{integer, frames per second (1 is recommended)}

\item{model}{\code{btvoccu} model object}

\item{x}{numeric covariate array}

\item{value}{"Occupancy", "Detection", or "Presence"}

\item{q}{quantile}

\item{shape}{\code{SpatialPolygonsDataFrame} object}

\item{cities}{\code{sf} object, named points}

\item{sites}{vector (in order of \code{shape} object)}

\item{season}{integer}

\item{palette}{color palette}

\item{legend.position}{see \code{tmap::tm_layout()}}

\item{title.position}{see \code{tmap::tm_layout()}}

\item{title.size}{see \code{tmap::tm_layout()}}

\item{legend.title.size}{see \code{tmap::tm_layout()}}

\item{legend.text.size}{see \code{tmap::tm_layout()}}

\item{xlab.size}{see \code{tmap::tm_xlab()}}

\item{ylab.size}{see \code{tmap::tm_ylab()}}

\item{dots.size}{see \code{tmap::tm_dots()}}

\item{cities.text}{see \code{tmap::tm_text()}}

\item{cities.just}{see \code{tmap::tm_text()}}

\item{cities.size}{see \code{tmap::tm_text()}}

\item{cities.col}{see \code{tmap::tm_text()}}

\item{periods}{vector}

\item{compass.position}{see \code{tmap::tm_compass()}}
}
\value{
movie of \code{tmap} objects
}
\description{
Create thematic movie demonstrating how
output sigmoids vary across space.
}
