% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{btvoccu_predict}
\alias{btvoccu_predict}
\title{\code{btvoccu} predictions}
\usage{
btvoccu_predict(
  model,
  x,
  sites,
  season,
  periods,
  value = "Occupancy",
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975),
  ndraws = 1000,
  spline = TRUE,
  spline.predict = FALSE,
  nknots = 5
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric array: predictors array}

\item{sites}{character vector}

\item{season}{integer}

\item{periods}{integer vector}

\item{value}{'Occupancy', 'Detection', or 'Presence'}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}

\item{credible}{quantiles (length \code{3})}

\item{ndraws}{number of posterior draws}

\item{spline}{logical: default \code{TRUE} to fit smoothing splines}

\item{spline.predict}{logical: default \code{FALSE} to not predict missing values using splines}

\item{nknots}{see \code{smooth.spline()}}
}
\value{
data frame
}
\description{
Predict occupancy, detection, and presence
probabilities with uncertainty over time.
}
