% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_chains.R
\name{combine_chains}
\alias{combine_chains}
\title{Combine MCMC Chains}
\usage{
combine_chains(model, effect, mnridx, burnin = 0.5, thin = 1)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{effect}{"betas", "alphas", or "thetas"}

\item{mnridx}{integer index}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
vector of posterior samples
}
\description{
Combine chains for mixing diagnostics
and posterior inference.
}
