% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covariate.R
\name{plot_covariate}
\alias{plot_covariate}
\title{Covariate time series plot}
\usage{
plot_covariate(
  x,
  sites,
  season,
  periods,
  covariate,
  yaxis,
  xaxis = "Period",
  legendtitle = "Site"
)
}
\arguments{
\item{x}{numeric array: covariate array}

\item{sites}{vector}

\item{season}{length \code{1} vector}

\item{periods}{vector}

\item{covariate}{character}

\item{yaxis}{characer: y-axis label}

\item{xaxis}{character: x-axis label}

\item{legendtile}{character: legend label}
}
\value{
\code{ggplot} object
}
\description{
Visualize covariates over time.
}
\examples{
plot_covariate(x, c("TOR", "YRK", "PEE"), 2005, 20:40, "popdensity", "Population Density")
plot_covariate(x, c("TOR", "YRK", "PEE"), 2005, 20:40, "temperature", "Temperature (Celsius)")

}
