% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{posterior_corrs}
\alias{posterior_corrs}
\title{Posterior correlations}
\usage{
posterior_corrs(
  model,
  Sigmas,
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{Sigmas}{"Sigmabetas" or "Sigmaalphas"}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}

\item{credible}{quantiles}
}
\value{
matrix
}
\description{
Summarize posterior correlations for beta and alpha coefficients.
}
