% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{posterior_effects}
\alias{posterior_effects}
\title{Posterior effects}
\usage{
posterior_effects(
  model,
  burnin = 0.5,
  thin = 1,
  credible = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}

\item{credible}{quantiles}
}
\value{
matrix
}
\description{
Summarize posterior beta, alpha, and theta coefficients.
}
\examples{
posterior_effects(mdl, 0.10, 2, c(0.25, 0.5, 0.75))

}
