% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sigmoids.R
\name{posterior_sigmoids}
\alias{posterior_sigmoids}
\title{Posterior sigmoids}
\usage{
posterior_sigmoids(
  model,
  x,
  sites,
  seasons,
  periods,
  ndraws = 200,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric covariate array}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}

\item{ndraws}{number of posterior draws}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth draw}
}
\value{
numeric array
}
\description{
Compute posterior occupancy, detection,
and presence probabilities.
}
\examples{
arry <- posterior_simulate(mdl, x, c("TOR", "YRK", "PEE"), 2004:2005, 20:40)

}
