% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesanalysis.R
\name{rpresence}
\alias{rpresence}
\title{Simulate presence-absence data}
\usage{
rpresence(pre)
}
\arguments{
\item{pre}{numeric array: presence probabilities}
}
\value{
binary array
}
\description{
Simulate presence-absence data
based on presence probabilities.
}
