% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_4darray.R
\name{split_4darray}
\alias{split_4darray}
\title{Train-Test Split for 4D Array}
\usage{
split_4darray(y, props, sites = NULL, seasons = NULL, periods = NULL)
}
\arguments{
\item{y}{binary array: response array}

\item{props}{length \code{1} or \code{2} proportions}

\item{sites}{vector}

\item{seasons}{vector}

\item{periods}{vector}
}
\value{
list
}
\description{
Split 4D array into train,
test, and validate arrays.
Splitting is done by setting
values to NA because main function
\code{btvoccu()} ignores NA responses
when fitting.
}
