% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_4darray.R
\name{subset_4darray}
\alias{subset_4darray}
\title{Subset 4D Array}
\usage{
subset_4darray(x, d, labels)
}
\arguments{
\item{x}{numeric array}

\item{d}{integer: array dimension}

\item{labels}{character vector: predictor names}
}
\value{
numeric array
}
\description{
Select predictors for design arrays.
}
