% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_score.R
\name{waic_score}
\alias{waic_score}
\title{Watanabe Akaike Information Criterion}
\usage{
waic_score(
  model,
  x,
  y,
  sites = NULL,
  seasons = NULL,
  periods = NULL,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} output}

\item{x}{numeric array: predictors array}

\item{y}{binary array: response array}

\item{sites}{vector (default \code{NULL} uses training sites)}

\item{seasons}{vector (default \code{NULL} uses training seasons)}

\item{periods}{vector (default \code{NULL} uses training periods)}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{integer: default \code{1} implies no thinning, \code{2} means keep every other draw, etc.}
}
\value{
list
}
\description{
Compute Watanabe-Akaike Information Criterion.
When indices are set to \code{NULL},
WAIC is calculated using the training data.
}
