% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_score.R
\name{waic_score}
\alias{waic_score}
\title{Watanabe Akaike information criterion}
\usage{
waic_score(
  model,
  x,
  y,
  sites = NULL,
  seasons = NULL,
  periods = NULL,
  burnin = 0.5,
  thin = 1
)
}
\arguments{
\item{model}{\code{btvoccu} model object}

\item{x}{numeric array: covariate array}

\item{y}{binary array: response array}

\item{sites}{vector (default \code{NULL} uses training sites)}

\item{seasons}{vector (default \code{NULL} uses training seasons)}

\item{periods}{vector (default \code{NULL} uses training periods)}

\item{burnin}{percent of posterior samples to burn}

\item{thin}{keep every nth posterior draw}
}
\value{
list
}
\description{
Compute Watanabe-Akaike information criterion.
}
\examples{
waic_score(mdl, x, y, burnin = 0.10)
waic_score(mdl, x, y, c("TOR", "YRK", "PEE"), 2002:2005, 20:40)

}
